# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Suivi Académique",
        "description": "Documentation de l'API pour le suivi académique",
        "contact": {
            "email": "contact@suivi-academique.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur API"
        }
    ],
    "paths": {
...
        "/api/personnels": {
            "get": {
                "tags": [
                    "Personnel"
                ],
                "summary": "Obtenir la liste des personnels",
                "description": "Retourne la liste de tous les personnels",
                "operationId": "getPersonnelsList",
                "responses": {
                    "200": {
                        "description": "Liste récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "code_personnel": {
                                                "type": "string",
                                                "example": "PER-001"
                                            },
                                            "nom_personnel": {
                                                "type": "string",
                                                "example": "Doe"
                                            },
                                            "prenom_personnel": {
                                                "type": "string",
                                                "example": "John"
                                            },
                                            "sexe_personnel": {
                                                "type": "string",
                                                "enum": [
                                                    "M",
                                                    "F"
                                                ],
                                                "example": "M"
                                            },
                                            "phone_personnel": {
                                                "type": "string",
                                                "example": "699123456"
                                            },
                                            "login_personnel": {
                                                "type": "string",
                                                "example": "jdoe"
                                            },
                                            "type_personnel": {
                                                "type": "string",
                                                "enum": [
                                                    "ENSEIGNANT",
                                                    "RESPONSABLE_ACADEMIQUE",
                                                    "RESPONSABLE_DISCIPLINE"
                                                ],
                                                "example": "ENSEIGNANT"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Personnel"
                ],
                "summary": "Créer un personnel",
                "operationId": "storePersonnel",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "code_personnel": {
                                        "type": "string",
                                        "example": "PER-001"
                                    },
                                    "nom_personnel": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "prenom_personnel": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "sexe_personnel": {
                                        "type": "string",
                                        "enum": [
                                            "M",
                                            "F"
                                        ],
                                        "example": "M"
                                    },
                                    "phone_personnel": {
                                        "type": "string",
                                        "example": "699123456"
                                    },
                                    "login_personnel": {
                                        "type": "string",
                                        "example": "jdoe"
                                    },
                                    "password_personnel": {
                                        "type": "string",
                                        "example": "888rdqdf"
                                    },
                                    "type_personnel": {
                                        "type": "string",
                                        "enum": [
                                            "ENSEIGNANT",
                                            "RESPONSABLE_ACADEMIQUE",
                                            "RESPONSABLE_DISCIPLINE"
                                        ],
                                        "example": "ENSEIGNANT"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Personnel créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/personnels/{code_personnel}": {
            "get": {
                "tags": [
                    "Personnel"
                ],
                "summary": "Afficher un personnel",
                "operationId": "showPersonnel",
                "parameters": [
                    {
                        "name": "code_personnel",
                        "in": "path",
                        "description": "Code du personnel",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Personnel trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code_personnel": {
                                            "type": "string",
                                            "example": "PER-001"
                                        },
                                        "nom_personnel": {
                                            "type": "string",
                                            "example": "Doe"
                                        },
                                        "prenom_personnel": {
                                            "type": "string",
                                            "example": "John"
                                        },
                                        "sexe_personnel": {
                                            "type": "string",
                                            "enum": [
                                                "M",
                                                "F"
                                            ],
                                            "example": "M"
                                        },
                                        "phone_personnel": {
                                            "type": "string",
                                            "example": "699123456"
                                        },
                                        "login_personnel": {
                                            "type": "string",
                                            "example": "jdoe"
                                        },
                                        "type_personnel": {
                                            "type": "string",
                                            "enum": [
                                                "ENSEIGNANT",
                                                "RESPONSABLE_ACADEMIQUE",
                                                "RESPONSABLE_DISCIPLINE"
                                            ],
                                            "example": "ENSEIGNANT"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Personnel"
                ],
                "summary": "Modifier un personnel",
                "operationId": "updatePersonnel",
                "parameters": [
                    {
                        "name": "code_personnel",
                        "in": "path",
                        "description": "Code du personnel",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "code_personnel": {
                                        "type": "string",
                                        "example": "PER-001"
                                    },
                                    "nom_personnel": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "prenom_personnel": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "sexe_personnel": {
                                        "type": "string",
                                        "enum": [
                                            "M",
                                            "F"
                                        ],
                                        "example": "M"
                                    },
                                    "phone_personnel": {
                                        "type": "string",
                                        "example": "699123456"
                                    },
                                    "login_personnel": {
                                        "type": "string",
                                        "example": "jdoe"
                                    },
                                    "password_personnel": {
                                        "type": "string",
                                        "example": "888rdqdf"
                                    },
                                    "type_personnel": {
                                        "type": "string",
                                        "enum": [
                                            "ENSEIGNANT",
                                            "RESPONSABLE_ACADEMIQUE",
                                            "RESPONSABLE_DISCIPLINE"
                                        ],
                                        "example": "ENSEIGNANT"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Personnel mis à jour"
                    },
                    "404": {
                        "description": "Personnel non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Personnel"
                ],
                "summary": "Supprimer un personnel",
                "operationId": "deletePersonnel",
                "parameters": [
                    {
                        "name": "code_personnel",
                        "in": "path",
                        "description": "Code du personnel",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Supprimé avec succès"
                    }
                }
            }
        },
...
    },
...
}